Thu, Jul 21, 2022, 8:06 PM +00:00

+++ x-final/app.py
@@ -321,7 +321,7 @@ def gradebook():
             total = 0
             for (i,j) in zip(sgrade, sweight):
                 #print (i["grade"],j["weight"])
-                agrade = (i["grade"] * (j["weight"]/100))
+                agrade = int(i["grade"] * (j["weight"]/100))
                 #print(agrade)
                 total += agrade
                 print(total)
